refactor TransactionService into smaller services
- splited TransactionService into:
  - TransactionOperationService which contain:
    - BuyAsset
    - SellAsset
      - Splited BuyAsset and SellAsset into smaller private methods
  - TransactionHistoryService which contain:
    - getTransactionHistory
    - getTransactionHistoryByPortfolio
    - getAllTransactions
    - getTransactionsByUser
    - getTransactionById
    - mapTransactionsToDTO
    - getPortfolioByidAndUser
    - getTransactionsByPortfolio
  - TransactionSecurityService which contain:
    - markTransactionAsSuspicious
    - getSuspiciousTransactions
    - markHighValueTransaction
    - markFrequentTransactions
    - getTransactionDataForEmail
